+++ spec/helpers/spec-helper.js
@@ -32,3 +32,11 @@ function makeDate(year, month, day) {
         throw new Error("makeDate takes year, month, day");
     return new Date(Date.UTC(year, month, day, 0, 0, 0, 0));
 }
+
+// http://stackoverflow.com/questions/20068497/d3-transition-in-unit-testing
+function flushAllD3Transitions() {
+    var now = Date.now;
+    Date.now = function() { return Infinity; };
+    d3.timer.flush();
+    Date.now = now;
+}